+++ src/adaptor.rs
@@ -5,7 +5,7 @@ use crate::{errors::RealmError, value::Value};
 pub mod format;
 pub mod source;
 
-#[derive(Debug)]
+// #[derive(Debug)]
 pub struct Adaptor {
     source: Box<dyn Source>,
 }