+++ src/lib.rs
@@ -9,5 +9,6 @@ mod value;
 pub use adaptor::Adaptor;
 pub use errors::RealmError;
 use map::Map;
+pub use parser::Parser;
 pub use realm::Realm;
 pub use value::Value;